src/dataloader/pyproject.toml
[project]
name = "hackathon-dataloader"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "kreuzberg>=0.5",
    "google-genai>=1.0",
    "chromadb-client>=0.6",
    "neo4j>=5.26",
    "sqlalchemy[asyncio]>=2.0",
    "asyncpg>=0.30",
    "pydantic>=2.10",
    "pydantic-settings>=2.7",
    "httpx>=0.27",
    "structlog>=24.4",
    "python-dotenv>=1.0",
]

[dependency-groups]
dev = [
    "pytest>=8.0",
    "ruff>=0.8",
]

[tool.ruff]
line-length = 120
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP"]
